requirements.txt
psycopg2
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
gunicorn
pydantic[email]
python-multipart
python-jose